graph/requirements.txt
psycopg2-binary
faker
tabulate